+++ src/collector/CollectorWrapper.ts
@@ -1,11 +1,12 @@
 import { v4 as uuidv4 } from 'uuid'
-import ClickEventListener from '../event/listener/ClickEventListener'
+import PointerUpEventListener from '../event/listener/PointerUpEventListener'
 import { createSessionStartedEvent } from '../event/createSessionStartedEvent'
 import { CollectorOptions } from '../types'
 import ChangeEventListener from '../event/listener/ChangeEventListener'
 import EventHandler from '../event/handler/EventHandler'
 import UnloadEventListener from '../event/listener/UnloadEventListener'
 import { debugEventSessionSender, defaultEventSessionSender } from '../event/handler/eventSessionSender'
+import KeyUpEventListener from '../event/listener/KeyUpEventListener'
 
 class CollectorWrapper {
   readonly eventHandler: EventHandler
@@ -25,7 +26,8 @@ class CollectorWrapper {
   }
 
   private initializeEventHandlers() {
-    new ClickEventListener(this.eventHandler, this.window).init()
+    new PointerUpEventListener(this.eventHandler, this.window).init()
+    new KeyUpEventListener(this.eventHandler, this.window).init()
     new ChangeEventListener(this.eventHandler, this.window).init()
     new UnloadEventListener(this.eventHandler, this.window).init()
   }