+++ src/types.ts
@@ -2,7 +2,9 @@ export enum EventType {
   SessionStarted = 'sessionStarted',
   Unload = 'unload',
   Click = 'click',
+  PointerUp = 'pointerup',
   Change = 'change',
+  KeyUp = 'keyup',
 }
 
 export type TEvent = GravitySessionStartedEvent | GravityEvent | GravityCustomEvent | GravitySessionEndedEvent